dark-stories/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-genai==1.10.0
python-multipart==0.0.20
python-dotenv==1.0.1
